Update DrawerController.swift
Left, tight and centre view controllers may be set by the user in the willFinishLaunching with options. If this is the case, we do not need to reassign them in the decodeRestorableState method.

Double reassignment happens if:

1) User instantiates DrawerController with left and centre controller in the willFinishLaunchingWithOptions method
2) User returns all relevant view controllers in the viewControllerWithRestorationIdentifierPath method. Following this, decodeRestorableState is called on DrawerController and left and centre controllers are reassigned.

The proposed PR ensures that controllers are only assigned in the decodeRestorableState method if they are not already assigned by the user. Seems quite logical :)