Replace percentVisible with fractionVisible
Throughout the code 'percentVisible' is used to indicate the fraction
visible. The name is ambiguous as it may suggest that the value should
be from 0 to 100, whereas in fact it goes from 0 to 1 (maybe a bit more
with overshoot).

I therefore propose to rename percentVisible into fractionVisible to
make it more clear what this variable is for.